custom-set-variables with vanilla Emacs
Vanilla out-of-the-box Emacs seems to honour the warnings about not using
custom-set-variables outside of the one emacs likes to maintain in your
configuration file rather aggressively in a way that somehow spacemacs does
not (because `custom-file` is not set to `init.el`?  but somehow spacemacs also
seems to always be trying to edit my `init.el` for custom settings, just less
aggressively somehow).

So use customize-set-variable for the vanilla Emacs.